+++ src/components/Button/Button.tsx
@@ -1,82 +1,141 @@
-import * as React from 'react';
+import React from 'react';
+import { Spinner } from '../Spinner';
 
-type Variant = 'inverted' | 'default' | 'outlined';
-type Size = 'small' | 'medium' | 'large' | 'none';
-type Shape = 'flated' | 'rounded' | 'semi-rounded';
-type Color = 'primary' | 'success' | 'error' | 'secondary';
+type Color = 'primary' | 'danger' | 'success';
+type Variant = 'outlined' | 'inverted' | 'default';
 
-export const Button: React.FC<
-    React.DetailedHTMLProps<React.ButtonHTMLAttributes<HTMLButtonElement>, HTMLButtonElement> & {
-        variant?: Variant;
-        size?: Size;
-        shape?: Shape;
-        className?: string;
-        color?: Color;
-    }
-> = (props) => {
-    const { variant = 'default', children, size = 'medium', shape = 'rounded', className } = props;
+type Props = React.ButtonHTMLAttributes<HTMLButtonElement> & {
+    color?: Color;
+    variant?: Variant;
+    isLoading?: boolean;
+    loadingText?: string;
+};
 
-    const borderRadius = React.useMemo(() => {
-        if (shape === 'semi-rounded') return 'rounded';
-        if (shape === 'rounded') return 'rounded-full';
-        return '';
-    }, [shape]);
+const _colors = (color?: Color, variant?: Variant) => {
+    const _class =
+        color === 'danger'
+            ? `${
+                  variant === 'default'
+                      ? 'text-white bg-rose-500 hover:bg-rose-600 dark:text-rose-300 dark:bg-rose-700 dark:hover:bg-rose-800'
+                      : 'text-rose-600 bg-rose-50 hover:bg-rose-100 dark:text-rose-300 dark:bg-slate-700 dark:hover:bg-black/20'
+              }
+  ${variant === 'outlined' ? ' border-rose-400 dark:border-rose-300' : 'border-transparent'}`
+            : color === 'success'
+              ? `${
+                    variant === 'default'
+                        ? 'text-white bg-green-500 hover:bg-green-600 dark:text-green-300 dark:bg-green-700 dark:hover:bg-green-800'
+                        : 'text-green-600 bg-green-50 hover:bg-green-100 dark:text-green-300 dark:bg-slate-700 dark:hover:bg-black/20'
+                } 
+  ${variant === 'outlined' ? ' border-green-500 dark:border-green-400' : 'border-transparent'}`
+              : `${
+                    variant === 'default'
+                        ? 'text-white bg-primary hover:bg-primary-dark dark:text-sky-300 dark:bg-sky-700 dark:hover:bg-sky-800'
+                        : 'text-primary bg-sky-100/70 hover:bg-sky-100 dark:text-sky-300 dark:bg-slate-700 dark:hover:bg-black/20 '
+                }
+  ${variant === 'outlined' ? ' border-primary dark:border-sky-400' : 'border-transparent'}`;
 
-    let _className = `px-4 ${size === 'small' ? 'py-1' : size === 'large' ? 'py-3' : size === 'none' ? '' : 'py-2'}
- ${borderRadius} text-sm font-medium transition-all ease-in-out
- ${variant === 'outlined'
-            ? 'border border-primary text-primary dark:text-gray-200  bg-transparent hover:bg-primary disabled:border-gray-200 disabled:text-gray-300 hover:text-white disabled:hover:bg-transparent dark:hover:bg-black/10 dark:hover:text-gray-200 dark:disabled:border-gray-500 dark:disabled:text-gray-400 dark:disabled:hover:bg-transparent'
-            : ''
-        }
- ${variant === 'default'
-            ? ` text-white dark:text-gray-200 dark:hover:text-gray-200  dark:disabled:text-gray-400
- bg-primary  dark:bg-primary disabled:bg-gray-200 hover:bg-primary-dark  dark:disabled:bg-gray-500   `
-            : ''
-        }`;
+    return _class;
+};
 
-    if (props.className) _className = _className + ' ' + className;
+const _colorsText = (color?: Color) => {
+    const _class =
+        color === 'danger'
+            ? 'text-rose-600 dark:text-red-400 hover:bg-rose-50 hover:dark:bg-white/5 hover:dark:text-red-300'
+            : color === 'success'
+              ? 'text-green-600 dark:text-green-400 hover:bg-green-50 hover:dark:bg-white/5 hover:dark:text-green-300'
+              : 'text-primary hover:bg-sky-50 dark:text-sky-300 dark:hover:bg-slate-600';
+    return _class;
+};
+
+export const Button = ({ children, className, color = 'primary', variant = 'default', isLoading = false, loadingText = '', ...rest }: Props) => {
+    const _class = _colors(color, variant);
 
     return (
-        <>
-            <button {...props} className={_className}>
-                {children}
-            </button>
-        </>
+        <button {...rest} className={`flex items-center justify-center cursor-pointer transition-all px-4 py-2 rounded border ${_class} ${className}`}>
+            {isLoading ? (
+                <>
+                    <span className="pr-1">{loadingText}</span>
+                    <Spinner />
+                </>
+            ) : (
+                children
+            )}
+        </button>
     );
 };
 
-type LinkButtonProps = React.AnchorHTMLAttributes<HTMLAnchorElement> & {
-    color?: "primary" | "danger" | "success"
-    variant?: "outlined" | "contained"
-}
-export const LinkButton = ({ children, className, color = "primary", variant = "contained", ...rest }: LinkButtonProps) => {
-    const _class = color === 'danger' ? `${variant === "outlined" ? 'border-rose-600 text-rose-600 hover:bg-rose-600 hover:text-white dark:border-rose-400  dark:text-rose-400 dark:hover:bg-rose-500 dark:hover:text-white' : 'border-rose-600 text-white bg-rose-600 hover:bg-rose-700'}`
-        : color === 'success' ? `${variant === "outlined" ? 'border-green-600 text-green-600 hover:bg-green-600 hover:text-white dark:border-green-400  dark:text-green-400 dark:hover:bg-green-500 dark:hover:text-white' : 'border-green-600 text-white bg-green-600 hover:bg-green-700'}`
-            : `${variant === "outlined" ? 'border-primary dark:border-sky-400 text-primary dark:text-sky-400 ' : 'border-primary text-white bg-primary hover:bg-primary-dark'}`
+export const TextButton = ({ children, color = 'primary', isLoading = false, loadingText = 'Wait', className, ...props }: Props) => {
+    const _class = _colorsText(color);
+    return (
+        <button className={`flex items-center  justify-center py-1 px-3 transition-all ${_class} ${className}`} {...props}>
+            {isLoading ? (
+                <>
+                    <span className="pr-1">{loadingText}</span>
+                    <Spinner />
+                </>
+            ) : (
+                children
+            )}
+        </button>
+    );
+};
 
-    return <a {...rest} className={`flex items-center cursor-pointer  transition-all px-4 py-2 rounded border  ${_class} ${className}`}
-    >
-        {children}
-    </a>
-}
+export const LinkButton = ({
+    children,
+    className,
+    color = 'primary',
+    variant = 'default',
+    ...rest
+}: React.AnchorHTMLAttributes<HTMLAnchorElement> & {
+    color?: Color;
+    variant?: Variant;
+}) => {
+    const _class = _colors(color, variant);
 
+    return (
+        <a {...rest} className={`flex items-center justify-center cursor-pointer  transition-all px-4 py-2 rounded border ${_class} ${className}`}>
+            {children}
+        </a>
+    );
+};
+
+export const LinkTextButton = ({
+    children,
+    className,
+    color = 'primary',
+    ...rest
+}: React.AnchorHTMLAttributes<HTMLAnchorElement> & {
+    color?: Color;
+}) => {
+    const _class = _colorsText(color);
+    return (
+        <a {...rest} className={`flex items-center justify-center cursor-pointer  transition-all px-3 py-1 rounded  ${_class} ${className}`}>
+            {children}
+        </a>
+    );
+};
 
 export const IconButton: React.FC<
-    React.DetailedHTMLProps<React.ButtonHTMLAttributes<HTMLButtonElement>, HTMLButtonElement> & {
-        variant?: 'inverted' | 'contained';
+    React.ButtonHTMLAttributes<HTMLButtonElement> & {
+        variant?: 'inverted' | 'default' | 'contained' | 'filled';
         shape?: 'square' | 'rounded';
     }
 > = (props) => {
-    const { className = 'w-10 h-10', children, variant = 'inverted', shape = 'rounded' } = props;
+    const { className = 'w-10 h-10', children, variant = 'default', shape = 'rounded' } = props;
 
     let _className = `ring-0 outline-none flex items-center justify-center 
     cursor-pointer transition-all duration-300 ease-in-out 
     ${shape === 'rounded' ? 'rounded-full' : 'rounded'}
-    
-  ${variant === 'inverted'
-            ? 'text-white bg-white/10 hover:bg-white/20 '
-            : 'bg-white  text-gray-600 dark:text-white hover:text-primary dark:bg-white/30 hover:bg-white/60  dark:hover:bg-white/10 dark:hover:text-white'
-        }`;
+    ${
+        variant === 'contained'
+            ? 'bg-white hover:text-primary text-black  dark:bg-white/10 dark:hover:bg-white/20 dark:text-white'
+            : variant === 'inverted'
+              ? 'bg-black/10 hover:bg-black/20 text-black  dark:bg-white/10 dark:hover:bg-white/20 dark:text-white'
+              : variant === 'filled'
+                ? 'bg-white/10 hover:bg-white/20 text-white '
+                : 'text-black dark:text-white hover:bg-black/10 hover:dark:bg-white/10 '
+    }`;
+
     if (className) _className = _className + ' ' + className;
 
     return (