+++ src/components/Select/Select.tsx
@@ -15,9 +15,7 @@ export const Select: React.FC<
     );
 
     return (
-        <div
-            className={` rounded px-2 border dark:border-gray-500 focus-within:ring-1 bg-white dark:bg-black dark:text-gray-300 dark:bg-opacity-30 dark:focus-within:bg-opacity-10 dark:focus-within:text-gray-200   ${color}`}
-        >
+        <div className={` rounded px-2 border dark:border-gray-500 focus-within:ring-1 dark:text-gray-300  dark:focus-within:bg-white/10 dark:focus-within:text-gray-200   ${color}`}>
             <select {...rest} className={`w-full px-2 py-[0.6rem]  ring-0 focus:outline-none bg-transparent focus:ring-0 dark:bg-transparent dark:placeholder-gray-400` + inputClass}>
                 {children}
             </select>