+++ src/components/SelectField/SelectField.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useMemo, useState } from 'react';
+import React, { FC, useEffect, useMemo, useRef, useState } from 'react';
 
 export type SelectFieldOption =
     | {
@@ -7,7 +7,7 @@ export type SelectFieldOption =
       }
     | undefined;
 
-export const SelectField: FC<{
+type SelectFieldProp = {
     placeholder?: string;
     items: SelectFieldOption[];
     error?: boolean;
@@ -16,9 +16,13 @@ export const SelectField: FC<{
     searchPlaceholder?: string;
     onChange?: (value: any) => void;
     isReset?: boolean;
-}> = ({ placeholder = '', items = [], error = false, searchPlaceholder = 'search item', onChange, value }) => {
+    dropClass?: string;
+    drop?: 'top' | 'bottom' | 'top-right' | 'bottom-right';
+};
+export const SelectField: FC<SelectFieldProp> = ({ placeholder = '', items = [], error = false, searchPlaceholder = 'search item', onChange, value, dropClass = 'max-h-[600px]', drop = 'bottom' }) => {
     const [inputValue, setInputValue] = useState('');
     const [open, setOpen] = useState(false);
+    const divRef = useRef<HTMLDivElement>(null);
 
     const defaultSpace = useMemo(() => {
         if (placeholder === '') if (value?.toString() === '') return ' py-[0.73rem]';
@@ -34,20 +38,32 @@ export const SelectField: FC<{
         return placeholder;
     }, [value]);
 
-    // const filteredItems = useMemo(() => {
-    //   if(inputValue.length > 0) return items.filter(
-    //     item => item?.title.toLowerCase().startsWith(inputValue.toLowerCase())
-    //   )
-    //   return items
-    // }, [items, inputValue])
+    const filteredItems = useMemo(() => {
+        if (inputValue.length > 0) return items.filter((item) => item?.title.toLowerCase().startsWith(inputValue.toLowerCase()));
+        return items;
+    }, [items, inputValue]);
+
+    useEffect(() => {
+        function handleClickOutside(event: MouseEvent) {
+            if (divRef.current && !divRef.current.contains(event.target as Node)) {
+                setOpen(false);
+            }
+        }
+
+        document.addEventListener('mousedown', handleClickOutside);
+
+        return () => {
+            document.removeEventListener('mousedown', handleClickOutside);
+        };
+    }, []);
 
     return (
-        <div className="relative min-w-[180px] font-medium">
+        <div className="relative min-w-[180px] font-medium" ref={divRef}>
             <div
                 onClick={() => setOpen(!open)}
-                className={` w-full border select-none ${error && 'border-red-700 text-red-700'} bg-white  dark:bg-black/30 dark:text-gray-300 p-[0.4rem] flex items-center justify-between rounded  
+                className={` w-full border select-none ${error && 'border-red-700 text-red-700'} dark:text-gray-300 p-[0.4rem] flex items-center justify-between rounded  
         ${!value ? 'text-gray-500' : 'text-gray-500'}
-        ${open ? ` ring-1 ${error ? 'border-red-700 ring-red-700' : 'border-primary ring-primary'}` : 'border-gray-400 dark:border-gray-500'}`}
+        ${open ? ` dark:bg-white/10 ring-1 ${error ? 'border-red-700 ring-red-700' : 'border-primary ring-primary'}` : 'border-gray-400 dark:border-gray-500'}`}
             >
                 <span className={`${defaultSpace}`}>{normalizedTitle}</span>
                 <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className={`w-[14px] h-[14px] ${open && 'rotate-180'}`}>
@@ -55,12 +71,14 @@ export const SelectField: FC<{
                 </svg>
             </div>
             <ul
-                className={`absolute z-10 text-gray-500 max-h-[600px] bg-white dark:bg-slate-800 dark:text-gray-300 shadow drop-shadow-2xl  rounded  mt-1 overflow-y-auto transition-all duration-200   ${
+                className={`absolute z-50 text-gray-500  bg-white dark:bg-slate-700 dark:text-gray-300 shadow drop-shadow-2xl  rounded  mt-1 overflow-y-auto overflow-x-hidden transition-all duration-200  select-none  ${
                     // open ? "max-h-60  ring-accent ring-1" : "max-h-0"
                     open ? 'visible' : ' hidden'
-                } `}
+                } 
+                    ${drop === 'top' ? 'bottom-12' : drop === 'top-right' ? 'bottom-12 right-0' : drop === 'bottom-right' ? 'right-0' : ''}  
+                    ${dropClass}`}
             >
-                <div className="flex items-center px-2 sticky top-0 border-b dark:border-b-white/30 focus-within:border-primary bg-white dark:bg-gray-900  ">
+                <div className="mx-2 my-2 flex items-center px-2 sticky top-0 border rounded dark:border-white/20 focus-within:border-primary bg-white dark:bg-gray-900  ">
                     <svg viewBox="0 0 20 20" fill="currentColor" className="w-4 h-4  text-gray-500 dark:text-gray-400">
                         <path
                             fillRule="evenodd"
@@ -73,15 +91,15 @@ export const SelectField: FC<{
                         value={inputValue}
                         onChange={(e) => setInputValue(e.target.value.toLowerCase())}
                         placeholder={searchPlaceholder}
-                        className="placeholder:text-gray-400 p-2 outline-none font-thin placeholder:italic bg-transparent"
+                        className="flex-1 placeholder:text-gray-400 px-2 py-2 outline-none placeholder:font-light   bg-transparent"
                     />
                 </div>
-                {items?.map((item) => (
+                {filteredItems?.map((item) => (
                     <li
                         key={item?.title}
-                        className={` text-left p-2  cursor-default text-sm hover:bg-accent hover:text-white
-            ${item?.value.toString() === value?.toString() && 'bg-primary text-white'}
-            ${item?.title?.toLowerCase().startsWith(inputValue) ? 'block' : 'hidden'}`}
+                        className={` text-left p-2  cursor-default text-sm hover:bg-slate-50 dark:hover:bg-white/5 flex justify-between items-center
+            ${item?.value.toString() === value?.toString() && ' font-bold text-primary dark:text-sky-200'}
+            ${item?.title?.toLowerCase().startsWith(inputValue) ? 'block' : 'hidden'} `}
                         onClick={() => {
                             if (item?.value.toString() !== value.toString()) {
                                 if (onChange) onChange(item?.value);
@@ -90,7 +108,12 @@ export const SelectField: FC<{
                             }
                         }}
                     >
-                        {item?.title}
+                        <span className="flex-1 truncate">{item?.title}</span>
+                        {item?.value.toString() === value?.toString() && (
+                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" className="w-3 h-3 font-bold ">
+                                <path d="M21,7L9,19L3.5,13.5L4.91,12.09L9,16.17L19.59,5.59L21,7Z" fill="currentColor" />
+                            </svg>
+                        )}
                     </li>
                 ))}
             </ul>