+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 ## 1.1.0, Jan 24, 2024
 
+-   DateField: New
 -   TextField: review
+-   TextArea: review
 -   Modal: review
 -   Dialog: review
 -   CheckBox: review
--   DateField: New
 -   DropDown: review
 -   Button: review
+-   Tab: review
+-   SelectField: review
+-   SelectLabelField: review
+-   SearchField: review